requirements.txt
Jinja2==3.1.4
python-docx==0.8.11
PyMuPDF>=1.20.0,<2.0.0
MarkupSafe==3.0.2
resume-parser
google-generativeai
